+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Fixed
 
+- Break ordering problem (#385)
 - Documentation mismatch (#361)
 - Profile type not checked in input (#355)
 - Error reported by UBSan (#371)